requirements.txt
Flask==3.0.2
flask-cors==3.0.10
requests==2.26.0
beautifulsoup4==4.10.0
Jinja2==3.1.3
python-dotenv==0.19.1
